backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
psycopg2-binary==2.9.9
openai==1.3.0
numpy==1.26.0
requests==2.31.0
Werkzeug==2.3.7
python-dotenv==1.0.0 